Removed the typed event package
